Split files under acp module
