+++ crates/goose/src/acp/server/mod.rs
@@ -84,6 +84,14 @@ mod sessions;
 mod sources;
 mod tools;
 
+mod protocol;
+mod provider_config;
+mod session_types;
+mod tool_calls;
+pub use provider_config::*;
+pub use session_types::*;
+pub use tool_calls::*;
+
 pub type AcpProviderFactory = Arc<
     dyn Fn(
             String,
@@ -144,1048 +152,6 @@ async fn ensure_refresh_identity_current(
         .into_identity()?;
     if current_identity != *planned_identity {
         anyhow::bail!("provider inventory identity changed before refresh completed");
-    }
-
-    Ok(())
-}
-
-/// In-memory state for an active ACP session.
-///
-/// ## Terminology (temporary, until all clients migrate to ACP)
-///
-/// The ACP protocol uses "session" to mean the conversation as the human sees it —
-/// a durable, append-only exchange of messages. Internally, goose also has a concept
-/// called "Session" (the `sessions` DB table) which represents the agent's working
-/// state: the message list the LLM sees, compaction state, provider binding, etc.
-///
-/// The ACP session ID maps directly to a `sessions` row. The `sessions` HashMap
-/// below is keyed by session ID.
-struct GooseAcpSession {
-    agent: AgentHandle,
-    tool_requests: HashMap<String, crate::conversation::message::ToolRequest>,
-    /// For each tool_call_id that belongs to a multi-tool chain (run of
-    /// consecutive ToolRequest blocks within one assistant message), the chain
-    /// it belongs to. Populated when the assistant message is processed.
-    /// Used by `handle_tool_response` to detect when a chain has fully
-    /// completed and fire a single LLM summary covering the run.
-    chain_membership: HashMap<String, Arc<ToolChain>>,
-    /// Set of tool_call_ids whose ToolResponse has already been processed.
-    /// Drives the "all responses present" check for chain completion.
-    responded_tool_ids: HashSet<String>,
-    /// Tool_call_ids of chains that have already had a summary task fired.
-    /// Idempotence guard so we summarize each chain at most once.
-    summarized_chains: HashSet<String>,
-    cancel_token: Option<CancellationToken>,
-    /// Working directory set while the agent was still loading.
-    /// Applied once the agent becomes ready.
-    pending_working_dir: Option<std::path::PathBuf>,
-}
-
-/// A run of consecutive ToolRequest blocks within one assistant message,
-/// tracked by [`GooseAcpSession::chain_membership`]. Used to drive a single
-/// LLM summary for the whole run once every step has a recorded ToolResponse.
-#[derive(Debug, Clone)]
-struct ToolChain {
-    /// Tool call ids in document order. Always `len() >= 2`.
-    ids: Vec<String>,
-    /// The message_id of the assistant message containing these tool calls.
-    /// Used to persist chain summaries back to the messages table.
-    message_id: String,
-}
-
-/// Progress stages signalled by the background agent setup task via the watch
-/// channel.  `ProviderReady` fires as soon as the provider (and goose-mode)
-/// are initialized — before extensions finish loading.  `FullyReady` fires
-/// once every extension has been loaded (or failed).
-#[derive(Clone)]
-enum AgentSetupProgress {
-    /// Provider is initialized; extensions are still loading in the background.
-    ProviderReady(Arc<Agent>),
-    /// Provider *and* all extensions are initialized.
-    FullyReady(Arc<Agent>),
-}
-
-type AgentSetupSignal = Option<Result<AgentSetupProgress, String>>;
-
-/// The agent may still be initializing in the background (extension loading,
-/// provider setup).  Callers that need the live agent (e.g. `on_prompt`) await
-/// the handle; callers that only need the session metadata can proceed without it.
-enum AgentHandle {
-    Ready(Arc<Agent>),
-    Loading(tokio::sync::watch::Receiver<AgentSetupSignal>),
-}
-
-struct AgentSetupRequest {
-    session_id: SessionId,
-    goose_session: Session,
-    mcp_servers: Vec<McpServer>,
-    /// Pre-resolved provider name + model config (from config, no network).
-    /// When present the spawn skips re-deriving these from config.
-    resolved_provider: Option<(String, crate::model::ModelConfig)>,
-    /// Pre-instantiated provider reused from synchronous session initialization.
-    prebuilt_provider: Option<Arc<dyn Provider>>,
-}
-
-pub struct GooseAcpAgentOptions {
-    pub provider_factory: AcpProviderFactory,
-    pub builtins: Vec<String>,
-    pub data_dir: std::path::PathBuf,
-    pub config_dir: std::path::PathBuf,
-    pub goose_mode: GooseMode,
-    pub disable_session_naming: bool,
-    pub goose_platform: GoosePlatform,
-    pub additional_source_roots: Vec<SourceRoot>,
-}
-
-pub struct GooseAcpAgent {
-    sessions: Arc<Mutex<HashMap<String, GooseAcpSession>>>,
-    provider_factory: AcpProviderFactory,
-    builtins: Vec<String>,
-    client_fs_capabilities: OnceCell<FileSystemCapabilities>,
-    client_terminal: OnceCell<bool>,
-    client_mcp_host_info: OnceCell<GooseMcpHostInfo>,
-    use_login_shell_path: OnceCell<bool>,
-    config_dir: std::path::PathBuf,
-    session_manager: Arc<SessionManager>,
-    permission_manager: Arc<PermissionManager>,
-    goose_mode: GooseMode,
-    disable_session_naming: bool,
-    provider_inventory: ProviderInventoryService,
-    goose_platform: GoosePlatform,
-    additional_source_roots: Vec<SourceRoot>,
-}
-
-/// Shorten a session/thread id for perf log correlation.
-/// All `perf:` logs use `sid=<8-char-prefix>` so a single session's activity
-/// can be extracted with `grep 'perf:' <log> | grep 'sid=abc12345'`.
-fn sid_short(id: &str) -> String {
-    id.chars().take(8).collect()
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct SessionListCursorToken {
-    updated_at: chrono::DateTime<chrono::Utc>,
-    // Goose stores updated_at with second precision in common write paths, so the
-    // cursor needs the full (updated_at, id) sort key to avoid skipping tied rows.
-    session_id: String,
-    filter_hash: String,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct SessionListCursorFilters {
-    cwd: Option<String>,
-    session_types: Vec<String>,
-    non_empty: bool,
-}
-
-fn invalid_session_list_cursor(message: &'static str) -> agent_client_protocol::Error {
-    agent_client_protocol::Error::invalid_params().data(message)
-}
-
-// bind cursors to the effective filters so they cannot be reused for a different list.
-fn session_list_filter_hash(
-    cwd: Option<&std::path::Path>,
-    session_types: &[SessionType],
-) -> Result<String, agent_client_protocol::Error> {
-    let mut session_type_names = session_types
-        .iter()
-        .map(ToString::to_string)
-        .collect::<Vec<_>>();
-    session_type_names.sort();
-    let filters = SessionListCursorFilters {
-        cwd: cwd.map(|path| path.to_string_lossy().to_string()),
-        session_types: session_type_names,
-        non_empty: true,
-    };
-    let bytes =
-        serde_json::to_vec(&filters).internal_err_ctx("Failed to encode session list filters")?;
-    Ok(URL_SAFE_NO_PAD.encode(Sha256::digest(bytes)))
-}
-
-fn decode_session_list_cursor(
-    cursor: Option<&str>,
-    cwd: Option<&std::path::Path>,
-    session_types: &[SessionType],
-) -> Result<Option<SessionListCursor>, agent_client_protocol::Error> {
-    let Some(cursor) = cursor else {
-        return Ok(None);
-    };
-
-    let bytes = URL_SAFE_NO_PAD
-        .decode(cursor)
-        .map_err(|_| invalid_session_list_cursor("malformed session list cursor"))?;
-    let token: SessionListCursorToken = serde_json::from_slice(&bytes)
-        .map_err(|_| invalid_session_list_cursor("malformed session list cursor"))?;
-
-    if token.session_id.is_empty() || token.filter_hash.is_empty() {
-        return Err(invalid_session_list_cursor("malformed session list cursor"));
-    }
-
-    let expected_filter_hash = session_list_filter_hash(cwd, session_types)?;
-    if token.filter_hash != expected_filter_hash {
-        return Err(invalid_session_list_cursor(
-            "session list cursor does not match filters",
-        ));
-    }
-
-    Ok(Some(SessionListCursor {
-        updated_at: token.updated_at,
-        session_id: token.session_id,
-    }))
-}
-
-fn encode_session_list_cursor(
-    cursor: &SessionListCursor,
-    cwd: Option<&std::path::Path>,
-    session_types: &[SessionType],
-) -> Result<String, agent_client_protocol::Error> {
-    let token = SessionListCursorToken {
-        updated_at: cursor.updated_at,
-        session_id: cursor.session_id.clone(),
-        filter_hash: session_list_filter_hash(cwd, session_types)?,
-    };
-    let bytes =
-        serde_json::to_vec(&token).internal_err_ctx("Failed to encode session list cursor")?;
-    Ok(URL_SAFE_NO_PAD.encode(bytes))
-}
-
-fn session_meta(session: &Session) -> serde_json::Map<String, serde_json::Value> {
-    let mut meta = serde_json::Map::new();
-    meta.insert(
-        "messageCount".to_string(),
-        serde_json::Value::Number(session.message_count.into()),
-    );
-    meta.insert(
-        "createdAt".to_string(),
-        serde_json::Value::String(session.created_at.to_rfc3339()),
-    );
-    if let Some(ref archived_at) = session.archived_at {
-        meta.insert(
-            "archivedAt".to_string(),
-            serde_json::Value::String(archived_at.to_rfc3339()),
-        );
-    }
-    meta.insert(
-        "userSetName".to_string(),
-        serde_json::Value::Bool(session.user_set_name),
-    );
-
-    if let Some(ref pid) = session.project_id {
-        meta.insert(
-            "projectId".to_string(),
-            serde_json::Value::String(pid.clone()),
-        );
-    }
-    if let Some(ref provider) = session.provider_name {
-        meta.insert(
-            "providerId".to_string(),
-            serde_json::Value::String(provider.clone()),
-        );
-    }
-    if let Some(ref mc) = session.model_config {
-        meta.insert(
-            "modelId".to_string(),
-            serde_json::Value::String(mc.model_name.clone()),
-        );
-    }
-    meta
-}
-
-fn spawn_session_name_update_notifier(
-    cx: ConnectionTo<Client>,
-) -> tokio::sync::mpsc::UnboundedSender<crate::session::SessionNameUpdate> {
-    let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel::<crate::session::SessionNameUpdate>();
-    tokio::spawn(async move {
-        while let Some(update) = rx.recv().await {
-            let mut meta = serde_json::Map::new();
-            meta.insert(
-                "messageCount".to_string(),
-                serde_json::Value::Number(update.message_count.into()),
-            );
-            meta.insert(
-                "userSetName".to_string(),
-                serde_json::Value::Bool(update.user_set_name),
-            );
-            let notification = SessionNotification::new(
-                SessionId::new(update.session_id.clone()),
-                SessionUpdate::SessionInfoUpdate(
-                    SessionInfoUpdate::new()
-                        .title(update.name)
-                        .updated_at(update.updated_at.to_rfc3339())
-                        .meta(meta),
-                ),
-            );
-            if let Err(error) = cx.send_notification(notification) {
-                warn!(
-                    session_id = %update.session_id,
-                    error = %error,
-                    "Failed to send generated session name update"
-                );
-            }
-        }
-    });
-    tx
-}
-
-fn extract_timeout_from_meta(meta: &Option<Meta>) -> Option<u64> {
-    meta.as_ref()
-        .and_then(|m| m.get("timeout"))
-        .and_then(|v| v.as_u64())
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct GooseClientMetaEnvelope {
-    #[serde(default)]
-    goose: Option<GooseClientMeta>,
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct GooseClientMeta {
-    #[serde(rename = "mcpHostCapabilities", default)]
-    mcp_host_capabilities: Option<GooseMcpHostCapabilities>,
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct GooseMcpHostCapabilities {
-    #[serde(default)]
-    extensions: Option<rmcp::model::ExtensionCapabilities>,
-}
-
-fn extract_goose_client_meta(meta: &Meta) -> Option<GooseClientMetaEnvelope> {
-    serde_json::from_value(serde_json::Value::Object(meta.clone())).ok()
-}
-
-fn extract_client_mcp_host_info(args: &InitializeRequest) -> GooseMcpHostInfo {
-    let host_capabilities = args
-        .client_capabilities
-        .meta
-        .as_ref()
-        .and_then(extract_goose_client_meta)
-        .and_then(|meta| meta.goose)
-        .and_then(|goose| goose.mcp_host_capabilities);
-    let explicit_extensions = host_capabilities
-        .as_ref()
-        .and_then(|capabilities| capabilities.extensions.as_ref())
-        .is_some();
-    let extensions = host_capabilities
-        .and_then(|capabilities| capabilities.extensions)
-        .unwrap_or_default();
-
-    GooseMcpHostInfo {
-        explicit_extensions,
-        extensions,
-        client_name: args.client_info.as_ref().map(|info| info.name.clone()),
-        client_version: args.client_info.as_ref().map(|info| info.version.clone()),
-    }
-}
-
-fn extract_use_login_shell_path(args: &InitializeRequest) -> bool {
-    args.meta
-        .as_ref()
-        .and_then(|meta| meta.get("goose/useLoginShellPath"))
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false)
-}
-
-fn mcp_server_to_extension_config(mcp_server: McpServer) -> Result<ExtensionConfig, String> {
-    match mcp_server {
-        McpServer::Stdio(stdio) => {
-            let timeout = extract_timeout_from_meta(&stdio.meta);
-            Ok(ExtensionConfig::Stdio {
-                name: stdio.name,
-                description: String::new(),
-                cmd: stdio.command.to_string_lossy().to_string(),
-                args: stdio.args,
-                envs: Envs::new(stdio.env.into_iter().map(|e| (e.name, e.value)).collect()),
-                env_keys: vec![],
-                timeout,
-                bundled: Some(false),
-                available_tools: vec![],
-            })
-        }
-        McpServer::Http(http) => {
-            let timeout = extract_timeout_from_meta(&http.meta);
-            Ok(ExtensionConfig::StreamableHttp {
-                name: http.name,
-                description: String::new(),
-                uri: http.url,
-                envs: Envs::default(),
-                env_keys: vec![],
-                headers: http
-                    .headers
-                    .into_iter()
-                    .map(|h| (h.name, h.value))
-                    .collect(),
-                timeout,
-                socket: None,
-                bundled: Some(false),
-                available_tools: vec![],
-            })
-        }
-        McpServer::Sse(_) => Err("SSE is unsupported, migrate to streamable_http".to_string()),
-        _ => Err("Unknown MCP server type".to_string()),
-    }
-}
-
-fn get_requested_line(arguments: Option<&rmcp::model::JsonObject>) -> Option<u32> {
-    arguments
-        .and_then(|args| args.get("line"))
-        .and_then(|v| v.as_u64())
-        .map(|l| l as u32)
-}
-
-fn is_developer_file_tool(tool_name: &str) -> bool {
-    matches!(tool_name, "read" | "write" | "edit")
-}
-
-fn extract_locations_from_meta(
-    tool_response: &crate::conversation::message::ToolResponse,
-) -> Option<Vec<ToolCallLocation>> {
-    let result = tool_response.tool_result.as_ref().ok()?;
-    let meta = result.meta.as_ref()?;
-    let locations_val = meta.get("tool_locations")?;
-    let entries: Vec<serde_json::Value> = serde_json::from_value(locations_val.clone()).ok()?;
-    let locations = entries
-        .into_iter()
-        .filter_map(|entry| {
-            let path = entry.get("path")?.as_str()?;
-            let line = entry.get("line").and_then(|v| v.as_u64()).map(|l| l as u32);
-            Some(ToolCallLocation::new(path).line(line))
-        })
-        .collect::<Vec<_>>();
-    if locations.is_empty() {
-        None
-    } else {
-        Some(locations)
-    }
-}
-
-fn extract_tool_locations(
-    tool_request: &crate::conversation::message::ToolRequest,
-    tool_response: &crate::conversation::message::ToolResponse,
-) -> Vec<ToolCallLocation> {
-    let mut locations = Vec::new();
-
-    if let Ok(tool_call) = &tool_request.tool_call {
-        if !is_developer_file_tool(tool_call.name.as_ref()) {
-            return locations;
-        }
-
-        let tool_name = tool_call.name.as_ref();
-        let path_str = tool_call
-            .arguments
-            .as_ref()
-            .and_then(|args| args.get("path"))
-            .and_then(|p| p.as_str());
-
-        if let Some(path_str) = path_str {
-            if matches!(tool_name, "read") {
-                let line = get_requested_line(tool_call.arguments.as_ref());
-                locations.push(ToolCallLocation::new(path_str).line(line));
-                return locations;
-            }
-
-            if matches!(tool_name, "write" | "edit") {
-                locations.push(ToolCallLocation::new(path_str).line(1));
-                return locations;
-            }
-
-            let command = tool_call
-                .arguments
-                .as_ref()
-                .and_then(|args| args.get("command"))
-                .and_then(|c| c.as_str());
-
-            if let Ok(result) = &tool_response.tool_result {
-                for content in &result.content {
-                    if let RawContent::Text(text_content) = &content.raw {
-                        let text = &text_content.text;
-
-                        match command {
-                            Some("view") => {
-                                let line = extract_view_line_range(text)
-                                    .map(|range| range.0 as u32)
-                                    .or(Some(1));
-                                locations.push(ToolCallLocation::new(path_str).line(line));
-                            }
-                            Some("str_replace") | Some("insert") => {
-                                let line = extract_first_line_number(text)
-                                    .map(|l| l as u32)
-                                    .or(Some(1));
-                                locations.push(ToolCallLocation::new(path_str).line(line));
-                            }
-                            Some("write") => {
-                                locations.push(ToolCallLocation::new(path_str).line(1));
-                            }
-                            _ => {
-                                locations.push(ToolCallLocation::new(path_str).line(1));
-                            }
-                        }
-                        break;
-                    }
-                }
-            }
-
-            if locations.is_empty() {
-                locations.push(ToolCallLocation::new(path_str).line(1));
-            }
-        }
-    }
-
-    locations
-}
-
-fn extract_view_line_range(text: &str) -> Option<(usize, usize)> {
-    let re = regex::Regex::new(r"\(lines (\d+)-(\d+|end)\)").ok()?;
-    if let Some(caps) = re.captures(text) {
-        let start = caps.get(1)?.as_str().parse::<usize>().ok()?;
-        let end = if caps.get(2)?.as_str() == "end" {
-            start
-        } else {
-            caps.get(2)?.as_str().parse::<usize>().ok()?
-        };
-        return Some((start, end));
-    }
-    None
-}
-
-fn extract_first_line_number(text: &str) -> Option<usize> {
-    let re = regex::Regex::new(r"```[^\n]*\n(\d+):").ok()?;
-    if let Some(caps) = re.captures(text) {
-        return caps.get(1)?.as_str().parse::<usize>().ok();
-    }
-    None
-}
-
-fn read_resource_link(link: ResourceLink) -> Option<String> {
-    let url = Url::parse(&link.uri).ok()?;
-    if url.scheme() == "file" {
-        let path = url.to_file_path().ok()?;
-        let contents = fs::read_to_string(&path).ok()?;
-
-        Some(format!(
-            "\n\n# {}\n```\n{}\n```",
-            path.to_string_lossy(),
-            contents
-        ))
-    } else {
-        None
-    }
-}
-
-fn format_tool_name(tool_name: &str) -> String {
-    if let Some((extension, tool)) = tool_name.split_once("__") {
-        format!(
-            "{}: {}",
-            extension.replace('_', " "),
-            tool.replace('_', " ")
-        )
-    } else {
-        tool_name.replace('_', " ")
-    }
-}
-
-/// Build a short fallback title from the tool name and arguments by extracting
-/// the most useful value (file path, command, query, url, etc.).
-fn summarize_tool_call(tool_name: &str, arguments: Option<&serde_json::Value>) -> String {
-    let base = format_tool_name(tool_name);
-
-    let detail = arguments.and_then(|args| {
-        let obj = args.as_object()?;
-        let keys = [
-            "path", "file", "command", "query", "url", "uri", "name", "pattern", "source",
-        ];
-        for key in &keys {
-            if let Some(v) = obj.get(*key) {
-                let s = match v {
-                    serde_json::Value::String(s) => s.clone(),
-                    other => other.to_string(),
-                };
-                if !s.is_empty() {
-                    let first_line = s.lines().next().unwrap_or(&s);
-                    if first_line.len() > 60 {
-                        return Some(format!("{}…", crate::utils::safe_truncate(first_line, 57)));
-                    }
-                    return Some(first_line.to_string());
-                }
-            }
-        }
-        None
-    });
-
-    match detail {
-        Some(d) => format!("{base} · {d}"),
-        None => base,
-    }
-}
-
-fn tool_call_identity_meta(tool_request: &ToolRequest) -> Option<Meta> {
-    let tool_call = tool_request.tool_call.as_ref().ok()?;
-    let tool_name = tool_call.name.to_string();
-    let extension_name = tool_request
-        .tool_meta
-        .as_ref()
-        .and_then(|meta| meta.get("goose_extension"))
-        .and_then(serde_json::Value::as_str)
-        .map(ToString::to_string)
-        .or_else(|| {
-            tool_name
-                .split_once("__")
-                .map(|(extension_name, _)| extension_name.to_string())
-        });
-
-    let mut tool_call_meta = serde_json::Map::new();
-    tool_call_meta.insert("toolName".to_string(), serde_json::Value::String(tool_name));
-    if let Some(extension_name) = extension_name {
-        tool_call_meta.insert(
-            "extensionName".to_string(),
-            serde_json::Value::String(extension_name),
-        );
-    }
-
-    let mut goose_meta = serde_json::Map::new();
-    goose_meta.insert(
-        "toolCall".to_string(),
-        serde_json::Value::Object(tool_call_meta),
-    );
-
-    let mut meta = serde_json::Map::new();
-    meta.insert("goose".to_string(), serde_json::Value::Object(goose_meta));
-    Some(meta)
-}
-
-/// Add `goose.toolChainSummary = { summary, count }` to a `Meta` blob,
-/// preserving any existing `goose.*` keys (e.g. `goose.toolCall` set by
-/// [`tool_call_identity_meta`]).
-fn with_tool_chain_summary_meta(base: Option<Meta>, summary: &str, count: usize) -> Option<Meta> {
-    let mut meta = base.unwrap_or_default();
-    let goose_entry = meta
-        .entry("goose".to_string())
-        .or_insert_with(|| serde_json::Value::Object(serde_json::Map::new()));
-    let goose_obj = match goose_entry {
-        serde_json::Value::Object(obj) => obj,
-        other => {
-            *other = serde_json::Value::Object(serde_json::Map::new());
-            match other {
-                serde_json::Value::Object(obj) => obj,
-                _ => unreachable!(),
-            }
-        }
-    };
-    let mut chain = serde_json::Map::new();
-    chain.insert(
-        "summary".to_string(),
-        serde_json::Value::String(summary.to_string()),
-    );
-    chain.insert(
-        "count".to_string(),
-        serde_json::Value::Number(serde_json::Number::from(count)),
-    );
-    goose_obj.insert(
-        "toolChainSummary".to_string(),
-        serde_json::Value::Object(chain),
-    );
-    Some(meta)
-}
-
-struct PendingToolCall {
-    tool_call: ToolCall,
-    identity_meta: Option<Meta>,
-    fallback_title: String,
-}
-
-/// Extract chains (runs of consecutive `MessageContent::ToolRequest` blocks)
-/// from a single message's content. Mirrors the frontend's chain detection in
-/// `MessageBubble.groupContentSections`: any non-tool block (text, thinking,
-/// image, etc.) breaks the run.
-///
-/// Returns one inner Vec per detected chain, holding the tool_call_ids in
-/// document order. Single-tool runs are included; callers (chain
-/// summarization) gate on `chain.len() >= 2`.
-///
-/// Note: this is the per-message view, kept around for tests and potential
-/// replay use. The live runtime path uses a streaming buffer fed by
-/// [`register_chain_buffer`] so chains that span multiple `AgentEvent::Message`
-/// events (e.g. Bedrock-style streaming, where one LLM message is split across
-/// rows — see `f087fa63c`) are still detected.
-#[allow(dead_code)]
-fn extract_tool_chains(
-    content: &[crate::conversation::message::MessageContent],
-) -> Vec<Vec<String>> {
-    use crate::conversation::message::MessageContent;
-    let mut chains: Vec<Vec<String>> = Vec::new();
-    let mut current: Vec<String> = Vec::new();
-
-    for block in content {
-        match block {
-            MessageContent::ToolRequest(tr) => current.push(tr.id.clone()),
-            MessageContent::ToolResponse(_) => {
-                // Server-side, assistant messages don't carry responses;
-                // responses arrive in subsequent messages. Treat as
-                // chain-neutral so a stray response doesn't split a chain
-                // if the data shape ever changes.
-            }
-            _ => {
-                if !current.is_empty() {
-                    chains.push(std::mem::take(&mut current));
-                }
-            }
-        }
-    }
-    if !current.is_empty() {
-        chains.push(current);
-    }
-    chains
-}
-
-/// If `buffer` holds a multi-tool run (≥ 2 tool requests), (re)register a
-/// [`ToolChain`] in `chain_membership` anchored on the **first** tool's
-/// message_id (the row [`SessionManager::update_tool_request_meta`] will patch
-/// when persisting the LLM-generated summary). Does **not** clear the buffer
-/// — chains can grow as more tools arrive (sequential tool use), so callers
-/// keep accumulating and re-registering with the larger set of ids.
-///
-/// The buffer contains `(tool_call_id, message_id)` pairs in arrival order,
-/// fed by the prompt stream loop. Sequential tool use (Bedrock/Anthropic)
-/// interleaves request → response → request → response across separate
-/// `AgentEvent::Message` events, so per-event `extract_tool_chains` only
-/// sees length-1 chains and would miss the run. Tool responses are
-/// chain-neutral (they don't split the run); only non-tool content (text,
-/// thinking, image, etc.) does, matching the frontend's
-/// `groupContentSections` behavior.
-fn extend_chain_membership(
-    buffer: &[(String, String)],
-    chain_membership: &mut HashMap<String, Arc<ToolChain>>,
-) {
-    if buffer.len() >= 2 {
-        let ids: Vec<String> = buffer.iter().map(|(id, _)| id.clone()).collect();
-        let message_id = buffer[0].1.clone();
-        let chain = Arc::new(ToolChain {
-            ids: ids.clone(),
-            message_id,
-        });
-        for id in ids {
-            chain_membership.insert(id, chain.clone());
-        }
-    }
-}
-
-fn pending_tool_call_from_request(tool_request: &ToolRequest) -> PendingToolCall {
-    let tool_name = match &tool_request.tool_call {
-        Ok(tool_call) => tool_call.name.to_string(),
-        Err(_) => "error".to_string(),
-    };
-    let args_value = tool_request
-        .tool_call
-        .as_ref()
-        .ok()
-        .and_then(|tc| tc.arguments.as_ref())
-        .map(|a| serde_json::Value::Object(a.clone()));
-    let fallback_title = summarize_tool_call(&tool_name, args_value.as_ref());
-    let identity_meta = tool_call_identity_meta(tool_request);
-
-    // Prefer the persisted LLM-generated title when available so replay (and
-    // any subsequent live initial ToolCall after the title task has already
-    // resolved) emits the nice title up front, with no flash of the
-    // deterministic fallback.
-    let initial_title = tool_request
-        .persisted_title()
-        .map(|s| s.to_string())
-        .unwrap_or_else(|| fallback_title.clone());
-
-    let mut tool_call = ToolCall::new(ToolCallId::new(tool_request.id.clone()), initial_title)
-        .status(ToolCallStatus::Pending);
-    if let Some(args) = args_value {
-        tool_call = tool_call.raw_input(args);
-    }
-
-    PendingToolCall {
-        tool_call,
-        identity_meta,
-        fallback_title,
-    }
-}
-
-fn builtin_to_extension_config(name: &str) -> ExtensionConfig {
-    if let Some(def) = PLATFORM_EXTENSIONS.get(name) {
-        ExtensionConfig::Platform {
-            name: def.name.into(),
-            description: def.description.into(),
-            display_name: Some(def.display_name.into()),
-            bundled: Some(true),
-            available_tools: vec![],
-        }
-    } else {
-        ExtensionConfig::Builtin {
-            name: name.into(),
-            display_name: None,
-            timeout: None,
-            bundled: Some(true),
-            description: name.into(),
-            available_tools: vec![],
-        }
-    }
-}
-
-fn build_model_state(current_model: &str, inventory: &ProviderInventoryEntry) -> SessionModelState {
-    let mut available_models = inventory
-        .models
-        .iter()
-        .map(|model| ModelInfo::new(ModelId::new(model.id.as_str()), model.name.as_str()))
-        .collect::<Vec<_>>();
-    if !available_models
-        .iter()
-        .any(|model| model.model_id.0.as_ref() == current_model)
-    {
-        available_models.insert(
-            0,
-            ModelInfo::new(ModelId::new(current_model), current_model),
-        );
-    }
-    SessionModelState::new(ModelId::new(current_model), available_models)
-}
-
-struct ProviderOptionEntry {
-    id: String,
-    label: String,
-}
-
-async fn list_provider_entries(current_provider: Option<&str>) -> Vec<ProviderOptionEntry> {
-    let mut providers = crate::providers::providers()
-        .await
-        .into_iter()
-        .map(|(metadata, _)| ProviderOptionEntry {
-            id: metadata.name,
-            label: metadata.display_name,
-        })
-        .collect::<Vec<_>>();
-    providers.sort_by(|left, right| left.id.cmp(&right.id));
-    providers.dedup_by(|left, right| left.id == right.id);
-
-    if let Some(current_provider) = current_provider {
-        if current_provider != DEFAULT_PROVIDER_ID
-            && !providers
-                .iter()
-                .any(|provider| provider.id == current_provider)
-        {
-            providers.push(ProviderOptionEntry {
-                id: current_provider.to_string(),
-                label: current_provider.to_string(),
-            });
-            providers.sort_by(|left, right| left.id.cmp(&right.id));
-        }
-    }
-
-    let mut entries = Vec::with_capacity(providers.len() + 1);
-    entries.push(ProviderOptionEntry {
-        id: DEFAULT_PROVIDER_ID.to_string(),
-        label: DEFAULT_PROVIDER_LABEL.to_string(),
-    });
-    entries.extend(providers);
-    entries
-}
-
-async fn build_provider_options(current_provider: Option<&str>) -> Vec<SessionConfigSelectOption> {
-    list_provider_entries(current_provider)
-        .await
-        .into_iter()
-        .map(|provider| SessionConfigSelectOption::new(provider.id, provider.label))
-        .collect()
-}
-
-fn session_provider_selection(session: &Session) -> &str {
-    session
-        .provider_name
-        .as_deref()
-        .unwrap_or(DEFAULT_PROVIDER_ID)
-}
-
-/// Resolve the provider name and model config for a session from an
-/// already-loaded `Config`.
-async fn resolve_provider_and_model_from_config(
-    config: &Config,
-    goose_session: &Session,
-) -> Result<(String, crate::model::ModelConfig), String> {
-    let global_provider = config.get_goose_provider().ok();
-    let provider_override = goose_session
-        .provider_name
-        .as_deref()
-        .filter(|p| *p != DEFAULT_PROVIDER_ID);
-    let provider_name = provider_override
-        .map(ToOwned::to_owned)
-        .or_else(|| global_provider.clone())
-        .ok_or_else(|| "Missing provider".to_string())?;
-    let explicitly_switched =
-        provider_override.is_some() && provider_override != global_provider.as_deref();
-    let model_config = match &goose_session.model_config {
-        Some(mc) => mc.clone(),
-        None if explicitly_switched => {
-            let entry = crate::providers::get_from_registry(&provider_name)
-                .await
-                .map_err(|e| e.to_string())?;
-            let default_model = &entry.metadata().default_model;
-            crate::model::ModelConfig::new(default_model)
-                .map_err(|e| e.to_string())?
-                .with_canonical_limits(&provider_name)
-        }
-        None => {
-            let model_id = config.get_goose_model().map_err(|e| e.to_string())?;
-            crate::model::ModelConfig::new(&model_id)
-                .map_err(|e| e.to_string())?
-                .with_canonical_limits(&provider_name)
-        }
-    };
-    Ok((provider_name, model_config))
-}
-
-fn with_preserved_session_request_params(
-    mut model_config: crate::model::ModelConfig,
-    current_model_config: Option<&crate::model::ModelConfig>,
-    request_params: Option<HashMap<String, serde_json::Value>>,
-) -> crate::model::ModelConfig {
-    let has_model_effort = model_config
-        .request_params
-        .as_ref()
-        .and_then(|params| params.get("thinking_effort"))
-        .is_some();
-    if !has_model_effort {
-        if let Some(thinking_effort) = current_model_config
-            .and_then(|config| config.request_params.as_ref())
-            .and_then(|params| params.get("thinking_effort"))
-            .cloned()
-        {
-            model_config = model_config.with_merged_request_params(HashMap::from([(
-                "thinking_effort".into(),
-                thinking_effort,
-            )]));
-        }
-    }
-    if let Some(request_params) = request_params {
-        model_config = model_config.with_merged_request_params(request_params);
-    }
-    model_config
-}
-
-/// Convenience wrapper: reads config from disk, then resolves provider + model.
-/// Cheap enough to call from `on_new_session` (file + registry reads, no network).
-async fn resolve_provider_and_model(
-    config_dir: &std::path::Path,
-    goose_session: &Session,
-) -> Result<(String, crate::model::ModelConfig), String> {
-    let config =
-        Config::new(config_dir.join(CONFIG_YAML_NAME), "goose").map_err(|e| e.to_string())?;
-    resolve_provider_and_model_from_config(&config, goose_session).await
-}
-
-fn build_mode_state(
-    current_mode: GooseMode,
-) -> Result<SessionModeState, agent_client_protocol::Error> {
-    let mut available = Vec::with_capacity(GooseMode::VARIANTS.len());
-    for &name in GooseMode::VARIANTS {
-        let goose_mode: GooseMode = name.parse().map_err(|_| {
-            agent_client_protocol::Error::internal_error() // impossible but satisfy linters
-                .data(format!("Failed to parse GooseMode variant: {}", name))
-        })?;
-        let mut mode = SessionMode::new(SessionModeId::new(name), name);
-        mode.description = goose_mode.get_message().map(Into::into);
-        available.push(mode);
-    }
-    Ok(SessionModeState::new(
-        SessionModeId::new(current_mode.to_string()),
-        available,
-    ))
-}
-
-fn should_refresh_inventory_for_session_init(entry: &ProviderInventoryEntry) -> bool {
-    entry.configured
-        && entry.supports_refresh
-        && (entry.last_updated_at.is_none() || ProviderInventoryService::is_stale(entry))
-}
-
-async fn build_eager_config_from_inventory(
-    provider_name: &str,
-    current_model: &str,
-    inventory: &ProviderInventoryEntry,
-    mode_state: &SessionModeState,
-    goose_session: &Session,
-) -> (SessionModelState, Vec<SessionConfigOption>) {
-    let ms = build_model_state(current_model, inventory);
-    let provider_selection = session_provider_selection(goose_session);
-    let provider_options = build_provider_options(Some(provider_name)).await;
-    let config_options =
-        build_config_options(mode_state, &ms, provider_selection, provider_options);
-    (ms, config_options)
-}
-
-fn build_config_options(
-    mode_state: &SessionModeState,
-    model_state: &SessionModelState,
-    provider_selection: &str,
-    provider_options: Vec<SessionConfigSelectOption>,
-) -> Vec<SessionConfigOption> {
-    let mode_options: Vec<SessionConfigSelectOption> = mode_state
-        .available_modes
-        .iter()
-        .map(|m| {
-            SessionConfigSelectOption::new(m.id.0.clone(), m.name.clone())
-                .description(m.description.clone())
-        })
-        .collect();
-    let model_options: Vec<SessionConfigSelectOption> = model_state
-        .available_models
-        .iter()
-        .map(|m| SessionConfigSelectOption::new(m.model_id.0.clone(), m.name.clone()))
-        .collect();
-    vec![
-        SessionConfigOption::select(
-            "provider",
-            "Provider",
-            provider_selection.to_string(),
-            provider_options,
-        ),
-        SessionConfigOption::select(
-            "mode",
-            "Mode",
-            mode_state.current_mode_id.0.clone(),
-            mode_options,
-        )
-        .category(SessionConfigOptionCategory::Mode),
-        SessionConfigOption::select(
-            "model",
-            "Model",
-            model_state.current_model_id.0.clone(),
-            model_options,
-        )
-        .category(SessionConfigOptionCategory::Model),
-    ]
-}
-
-fn to_nonnegative_u64(value: Option<i32>) -> Option<u64> {
-    value.and_then(|v| u64::try_from(v).ok())
-}
-
-fn build_prompt_usage(session: &Session) -> Option<Usage> {
-    let total = to_nonnegative_u64(session.total_tokens)?;
-    let input = to_nonnegative_u64(session.input_tokens).unwrap_or(0);
-    let output = to_nonnegative_u64(session.output_tokens).unwrap_or(0);
-    Some(Usage::new(total, input, output))
-}
-
-fn build_usage_update(session: &Session, context_limit: usize) -> UsageUpdate {
-    let used = session.total_tokens.unwrap_or(0).max(0) as u64;
-    UsageUpdate::new(used, context_limit as u64)
-}
-
-fn validate_absolute_cwd(cwd: &Path) -> Result<(), agent_client_protocol::Error> {
-    if !cwd.is_absolute() {
-        return Err(
-            agent_client_protocol::Error::invalid_params().data("cwd must be an absolute path")
-        );
-    }
-
-    if !cwd.exists() || !cwd.is_dir() {
-        return Err(agent_client_protocol::Error::invalid_params().data("invalid directory path"));
     }
 
     Ok(())
@@ -2477,1122 +1443,6 @@ fn extract_tool_raw_output(tool_result: &ToolResult<CallToolResult>) -> Option<s
         .as_ref()
         .ok()
         .and_then(|result| result.structured_content.clone())
-}
-
-impl GooseAcpAgent {
-    async fn on_initialize(
-        &self,
-        args: InitializeRequest,
-    ) -> Result<InitializeResponse, agent_client_protocol::Error> {
-        debug!(?args, "initialize request");
-
-        let _ = self
-            .client_fs_capabilities
-            .set(args.client_capabilities.fs.clone());
-        let _ = self.client_terminal.set(args.client_capabilities.terminal);
-        let _ = self
-            .client_mcp_host_info
-            .set(extract_client_mcp_host_info(&args));
-        let _ = self
-            .use_login_shell_path
-            .set(extract_use_login_shell_path(&args));
-
-        let capabilities = AgentCapabilities::new()
-            .load_session(true)
-            .session_capabilities(
-                SessionCapabilities::new()
-                    .list(SessionListCapabilities::new())
-                    .close(SessionCloseCapabilities::new()),
-            )
-            .prompt_capabilities(
-                PromptCapabilities::new()
-                    .image(true)
-                    .audio(false)
-                    .embedded_context(true),
-            )
-            .mcp_capabilities(McpCapabilities::new().http(true));
-        Ok(InitializeResponse::new(args.protocol_version)
-            .agent_capabilities(capabilities)
-            .auth_methods(vec![AuthMethod::Agent(
-                AuthMethodAgent::new("goose-provider", "Configure Provider")
-                    .description("Run `goose configure` to set up your AI provider and API key"),
-            )]))
-    }
-
-    async fn on_new_session(
-        &self,
-        cx: &ConnectionTo<Client>,
-        args: NewSessionRequest,
-    ) -> Result<NewSessionResponse, agent_client_protocol::Error> {
-        debug!(?args, "new session request");
-        let t_start = std::time::Instant::now();
-        validate_absolute_cwd(&args.cwd)?;
-
-        let requested_provider = args
-            .meta
-            .as_ref()
-            .and_then(|m| m.get("provider"))
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-
-        let project_id = args
-            .meta
-            .as_ref()
-            .and_then(|m| m.get("projectId"))
-            .and_then(|v| v.as_str())
-            .map(|s| s.to_string());
-
-        // When _meta.client is set, the session is created by a known client
-        // (e.g. "goose" for the desktop app) and treated as a User session.
-        // Without it, sessions default to Acp for programmatic ACP clients.
-        let session_type = match args
-            .meta
-            .as_ref()
-            .and_then(|m| m.get("client"))
-            .and_then(|v| v.as_str())
-        {
-            Some(_) => SessionType::User,
-            None => SessionType::Acp,
-        };
-
-        let t0 = std::time::Instant::now();
-        let goose_session = self
-            .session_manager
-            .create_session(
-                args.cwd.clone(),
-                "New Chat".to_string(),
-                session_type,
-                self.goose_mode,
-            )
-            .await
-            .internal_err_ctx("Failed to create session")?;
-
-        let mut builder = self.session_manager.update(&goose_session.id);
-        if let Some(ref provider) = requested_provider {
-            builder = builder.provider_name(provider);
-        }
-        if let Some(pid) = project_id {
-            builder = builder.project_id(Some(pid));
-        }
-        builder
-            .apply()
-            .await
-            .internal_err_ctx("Failed to update session")?;
-
-        let goose_session = self
-            .session_manager
-            .get_session(&goose_session.id, false)
-            .await
-            .internal_err_ctx("Failed to reload session")?;
-
-        let session_id_str = goose_session.id.clone();
-        let sid = sid_short(&session_id_str);
-        debug!(target: "perf", sid = %sid, ms = t0.elapsed().as_millis() as u64, "perf: new_session create_session");
-
-        let (agent_tx, agent_rx) = tokio::sync::watch::channel::<AgentSetupSignal>(None);
-
-        let acp_session = GooseAcpSession {
-            agent: AgentHandle::Loading(agent_rx),
-            tool_requests: HashMap::new(),
-            chain_membership: HashMap::new(),
-            responded_tool_ids: HashSet::new(),
-            summarized_chains: HashSet::new(),
-            cancel_token: None,
-            pending_working_dir: None,
-        };
-        self.sessions
-            .lock()
-            .await
-            .insert(session_id_str.clone(), acp_session);
-
-        let mode_state = build_mode_state(self.goose_mode)?;
-
-        let resolved = resolve_provider_and_model(&self.config_dir, &goose_session).await;
-        let initial_usage_update = resolved
-            .as_ref()
-            .ok()
-            .map(|(_, mc)| build_usage_update(&goose_session, mc.context_limit()));
-        let acp_session_id = SessionId::new(session_id_str);
-        let (model_state, config_options, prebuilt_provider) = self
-            .prepare_session_init_config(&resolved, &mode_state, &goose_session)
-            .await;
-
-        let working_dir = goose_session.working_dir.clone();
-
-        self.spawn_agent_setup(
-            cx,
-            agent_tx,
-            AgentSetupRequest {
-                session_id: acp_session_id.clone(),
-                goose_session,
-                mcp_servers: args.mcp_servers,
-                resolved_provider: resolved.as_ref().ok().cloned(),
-                prebuilt_provider,
-            },
-        );
-
-        let mut response = NewSessionResponse::new(acp_session_id.clone()).modes(mode_state);
-        if let Some(ms) = model_state {
-            response = response.models(ms);
-        }
-        if let Some(co) = config_options {
-            response = response.config_options(co);
-        }
-        if let Some(usage_update) = initial_usage_update {
-            cx.send_notification(SessionNotification::new(
-                acp_session_id.clone(),
-                SessionUpdate::UsageUpdate(usage_update),
-            ))?;
-        }
-        Self::send_available_commands_update(cx, &acp_session_id, &working_dir)?;
-        debug!(
-            target: "perf",
-            sid = %sid,
-            ms = t_start.elapsed().as_millis() as u64,
-            "perf: new_session done (agent setup continues in background)"
-        );
-        Ok(response)
-    }
-
-    /// Look up the session and return the agent if already ready, or the watch
-    /// receiver if still loading.  Optionally sets a cancellation token on the
-    /// session (needed by `on_prompt`).
-    async fn get_agent_or_receiver(
-        &self,
-        session_id: &str,
-        cancel_token: Option<CancellationToken>,
-    ) -> Result<
-        Either<Arc<Agent>, tokio::sync::watch::Receiver<AgentSetupSignal>>,
-        agent_client_protocol::Error,
-    > {
-        let mut sessions = self.sessions.lock().await;
-        let session = sessions.get_mut(session_id).ok_or_else(|| {
-            agent_client_protocol::Error::resource_not_found(Some(session_id.to_string()))
-                .data(format!("Session not found: {}", session_id))
-        })?;
-        if let Some(token) = cancel_token {
-            session.cancel_token = Some(token);
-        }
-        match &session.agent {
-            AgentHandle::Ready(agent) => Ok(Either::Left(agent.clone())),
-            AgentHandle::Loading(rx) => Ok(Either::Right(rx.clone())),
-        }
-    }
-
-    /// Wait until the agent is **fully ready** (provider + all extensions).
-    /// Most callers (e.g. `on_prompt`, `on_get_tools`) should use this.
-    async fn get_session_agent(
-        &self,
-        session_id: &str,
-        cancel_token: Option<CancellationToken>,
-    ) -> Result<Arc<Agent>, agent_client_protocol::Error> {
-        let mut rx = match self.get_agent_or_receiver(session_id, cancel_token).await? {
-            Either::Left(agent) => return Ok(agent),
-            Either::Right(rx) => rx,
-        };
-        // Wait specifically for FullyReady (not just ProviderReady).
-        let guard = rx
-            .wait_for(|v| {
-                matches!(
-                    v,
-                    Some(Ok(AgentSetupProgress::FullyReady(_))) | Some(Err(_))
-                )
-            })
-            .await
-            .map_err(|_| {
-                agent_client_protocol::Error::internal_error()
-                    .data("Agent setup task was dropped".to_string())
-            })?;
-        match guard.as_ref().unwrap() {
-            Ok(AgentSetupProgress::FullyReady(agent)) => Ok(agent.clone()),
-            Err(e) => Err(agent_client_protocol::Error::internal_error().data(e.clone())),
-            // wait_for predicate excludes ProviderReady
-            _ => unreachable!(),
-        }
-    }
-
-    /// Wait only until the **provider** is initialized.  Extensions may still
-    /// be loading in the background.  Use this for operations that only touch
-    /// the provider (e.g. `update_provider`, `set_model`, `build_config_update`).
-    async fn get_session_agent_provider_ready(
-        &self,
-        session_id: &str,
-    ) -> Result<Arc<Agent>, agent_client_protocol::Error> {
-        let mut rx = match self.get_agent_or_receiver(session_id, None).await? {
-            Either::Left(agent) => return Ok(agent),
-            Either::Right(rx) => rx,
-        };
-        // Any signal (ProviderReady, FullyReady, or Err) unblocks us.
-        let guard = rx.wait_for(|v| v.is_some()).await.map_err(|_| {
-            agent_client_protocol::Error::internal_error()
-                .data("Agent setup task was dropped".to_string())
-        })?;
-        match guard.as_ref().unwrap() {
-            Ok(progress) => match progress {
-                AgentSetupProgress::ProviderReady(agent)
-                | AgentSetupProgress::FullyReady(agent) => Ok(agent.clone()),
-            },
-            Err(e) => Err(agent_client_protocol::Error::internal_error().data(e.clone())),
-        }
-    }
-
-    async fn add_mcp_extensions(
-        agent: &Arc<Agent>,
-        mcp_servers: Vec<McpServer>,
-        session_id: &str,
-    ) -> Result<(), agent_client_protocol::Error> {
-        let mut configs = Vec::with_capacity(mcp_servers.len());
-        for mcp_server in mcp_servers {
-            let config = match mcp_server_to_extension_config(mcp_server) {
-                Ok(c) => c,
-                Err(msg) => {
-                    return Err(agent_client_protocol::Error::invalid_params().data(msg));
-                }
-            };
-            configs.push(config);
-        }
-
-        if configs.is_empty() {
-            return Ok(());
-        }
-
-        let results = agent
-            .add_extensions_bulk(configs, session_id)
-            .await
-            .internal_err()?;
-        for result in &results {
-            if !result.success {
-                let error_msg = result.error.as_deref().unwrap_or("unknown error");
-                return Err(agent_client_protocol::Error::internal_error().data(format!(
-                    "Failed to add MCP server '{}': {}",
-                    result.name, error_msg
-                )));
-            }
-        }
-        Ok(())
-    }
-
-    async fn on_load_session(
-        &self,
-        cx: &ConnectionTo<Client>,
-        args: LoadSessionRequest,
-    ) -> Result<LoadSessionResponse, agent_client_protocol::Error> {
-        debug!(?args, "load session request");
-        validate_absolute_cwd(&args.cwd)?;
-
-        let session_id = args.session_id.0.to_string();
-        let sid = sid_short(&session_id);
-        let t_start = std::time::Instant::now();
-
-        let t0 = std::time::Instant::now();
-        let goose_session = self
-            .session_manager
-            .get_session(&session_id, true)
-            .await
-            .map_err(|_| {
-                agent_client_protocol::Error::resource_not_found(Some(session_id.clone()))
-                    .data(format!("Session not found: {}", session_id))
-            })?;
-        debug!(target: "perf", sid = %sid, ms = t0.elapsed().as_millis() as u64, "perf: load_session get_session");
-        let loaded_mode = goose_session.goose_mode;
-
-        // ── REPLAY MESSAGES ──
-        // Stream user-visible messages back to the client so the chat view
-        // populates immediately, before the slow agent/provider/extension setup.
-        let messages = goose_session
-            .conversation
-            .as_ref()
-            .map(|c| c.messages().to_vec())
-            .unwrap_or_default();
-        debug!(
-            target: "perf",
-            sid = %sid,
-            messages = messages.len(),
-            "perf: load_session messages loaded"
-        );
-
-        let mut replay_tool_requests =
-            HashMap::<String, crate::conversation::message::ToolRequest>::new();
-
-        for message in &messages {
-            if !message.metadata.user_visible {
-                continue;
-            }
-
-            for content_item in &message.content {
-                match content_item {
-                    MessageContent::Text(text) => {
-                        let mut tc = TextContent::new(text.text.clone());
-                        if let Some(audience) = text.audience() {
-                            tc = tc.annotations(
-                                Annotations::new().audience(
-                                    audience
-                                        .iter()
-                                        .map(|r| match r {
-                                            Role::Assistant => {
-                                                agent_client_protocol::schema::Role::Assistant
-                                            }
-                                            Role::User => agent_client_protocol::schema::Role::User,
-                                        })
-                                        .collect::<Vec<_>>(),
-                                ),
-                            );
-                        }
-                        let chunk = ContentChunk::new(ContentBlock::Text(tc))
-                            .meta(replay_message_meta(message));
-                        let update = match message.role {
-                            Role::User => SessionUpdate::UserMessageChunk(chunk),
-                            Role::Assistant => SessionUpdate::AgentMessageChunk(chunk),
-                        };
-                        cx.send_notification(SessionNotification::new(
-                            args.session_id.clone(),
-                            update,
-                        ))?;
-                    }
-                    MessageContent::ToolRequest(tool_request) => {
-                        // Replay-only: emit the ToolCall notification and
-                        // stash the request for location extraction, but
-                        // don't require a full GooseAcpSession.
-                        replay_tool_requests.insert(tool_request.id.clone(), tool_request.clone());
-
-                        let pending_tool_call = pending_tool_call_from_request(tool_request);
-                        let mut meta = pending_tool_call.identity_meta;
-                        // If this tool request is the first of a chain whose
-                        // summary was persisted at completion time, attach the
-                        // chain summary to the initial ToolCall so the chain
-                        // header is correct on first paint after reload.
-                        if let Some(chain_summary) = tool_request.persisted_chain_summary() {
-                            meta = with_tool_chain_summary_meta(
-                                meta,
-                                &chain_summary.summary,
-                                chain_summary.count,
-                            );
-                        }
-                        let tool_call = pending_tool_call
-                            .tool_call
-                            .meta(merge_replay_message_meta(meta, message));
-
-                        cx.send_notification(SessionNotification::new(
-                            args.session_id.clone(),
-                            SessionUpdate::ToolCall(tool_call),
-                        ))?;
-                    }
-                    MessageContent::ToolResponse(tool_response) => {
-                        // Replay-only: emit the ToolCallUpdate notification,
-                        // using the stashed replay_tool_requests for location
-                        // extraction.
-                        let status = match &tool_response.tool_result {
-                            Ok(result) if result.is_error == Some(true) => ToolCallStatus::Failed,
-                            Ok(_) => ToolCallStatus::Completed,
-                            Err(_) => ToolCallStatus::Failed,
-                        };
-
-                        let mut fields = ToolCallUpdateFields::new().status(status);
-                        if let Some(raw_output) =
-                            extract_tool_raw_output(&tool_response.tool_result)
-                        {
-                            fields = fields.raw_output(raw_output);
-                        }
-                        if !tool_response
-                            .tool_result
-                            .as_ref()
-                            .is_ok_and(|r| r.is_acp_aware())
-                        {
-                            let content = build_tool_call_content(&tool_response.tool_result);
-                            fields = fields.content(content);
-
-                            let locations = extract_locations_from_meta(tool_response)
-                                .unwrap_or_else(|| {
-                                    if let Some(tool_request) =
-                                        replay_tool_requests.get(&tool_response.id)
-                                    {
-                                        extract_tool_locations(tool_request, tool_response)
-                                    } else {
-                                        Vec::new()
-                                    }
-                                });
-                            if !locations.is_empty() {
-                                fields = fields.locations(locations);
-                            }
-                        }
-
-                        let update =
-                            ToolCallUpdate::new(ToolCallId::new(tool_response.id.clone()), fields)
-                                .meta(merge_replay_message_meta(
-                                    extract_tool_call_update_meta(tool_response),
-                                    message,
-                                ));
-                        cx.send_notification(SessionNotification::new(
-                            args.session_id.clone(),
-                            SessionUpdate::ToolCallUpdate(update),
-                        ))?;
-                    }
-                    MessageContent::Thinking(thinking) => {
-                        cx.send_notification(SessionNotification::new(
-                            args.session_id.clone(),
-                            SessionUpdate::AgentThoughtChunk(
-                                ContentChunk::new(ContentBlock::Text(TextContent::new(
-                                    thinking.thinking.clone(),
-                                )))
-                                .meta(replay_message_meta(message)),
-                            ),
-                        ))?;
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        // Update working directory.
-        self.session_manager
-            .update(&session_id)
-            .working_dir(args.cwd.clone())
-            .apply()
-            .await
-            .internal_err_ctx("Failed to update session working directory")?;
-        let goose_session = self
-            .session_manager
-            .get_session(&session_id, false)
-            .await
-            .internal_err_ctx("Failed to reload session")?;
-
-        // Register the session with a Loading handle.
-        let (agent_tx, agent_rx) = tokio::sync::watch::channel::<AgentSetupSignal>(None);
-
-        let acp_session = GooseAcpSession {
-            agent: AgentHandle::Loading(agent_rx),
-            tool_requests: replay_tool_requests,
-            chain_membership: HashMap::new(),
-            responded_tool_ids: HashSet::new(),
-            summarized_chains: HashSet::new(),
-            cancel_token: None,
-            pending_working_dir: None,
-        };
-        self.sessions
-            .lock()
-            .await
-            .insert(session_id.clone(), acp_session);
-
-        let mode_state = build_mode_state(loaded_mode)?;
-
-        let resolved = resolve_provider_and_model(&self.config_dir, &goose_session).await;
-        let initial_usage_update = resolved
-            .as_ref()
-            .ok()
-            .map(|(_, mc)| build_usage_update(&goose_session, mc.context_limit()))
-            .or_else(|| {
-                goose_session
-                    .model_config
-                    .as_ref()
-                    .map(|mc| build_usage_update(&goose_session, mc.context_limit()))
-            });
-        let (model_state, config_options, prebuilt_provider) = self
-            .prepare_session_init_config(&resolved, &mode_state, &goose_session)
-            .await;
-
-        self.spawn_agent_setup(
-            cx,
-            agent_tx,
-            AgentSetupRequest {
-                session_id: args.session_id.clone(),
-                goose_session,
-                mcp_servers: args.mcp_servers,
-                resolved_provider: None,
-                prebuilt_provider,
-            },
-        );
-
-        let mut response = LoadSessionResponse::new().modes(mode_state);
-        if let Some(ms) = model_state {
-            response = response.models(ms);
-        }
-        if let Some(co) = config_options {
-            response = response.config_options(co);
-        }
-        if let Some(usage_update) = initial_usage_update {
-            cx.send_notification(SessionNotification::new(
-                args.session_id.clone(),
-                SessionUpdate::UsageUpdate(usage_update),
-            ))?;
-        }
-        Self::send_available_commands_update(cx, &args.session_id, &args.cwd)?;
-        debug!(
-            target: "perf",
-            sid = %sid,
-            ms = t_start.elapsed().as_millis() as u64,
-            "perf: load_session done (agent setup continues in background)"
-        );
-        Ok(response)
-    }
-
-    async fn on_prompt(
-        &self,
-        cx: &ConnectionTo<Client>,
-        args: PromptRequest,
-    ) -> Result<PromptResponse, agent_client_protocol::Error> {
-        // The ACP session_id IS the thread ID.
-        let session_id = args.session_id.0.to_string();
-        let sid = sid_short(&session_id);
-        let t_start = std::time::Instant::now();
-
-        let cancel_token = CancellationToken::new();
-        let agent = self
-            .get_session_agent(&session_id, Some(cancel_token.clone()))
-            .await?;
-
-        let user_message = Self::convert_acp_prompt_to_message(&args.prompt);
-
-        let message_text = user_message.as_concat_text();
-        if let Some(parsed) = crate::agents::execute_commands::parse_slash_command(&message_text) {
-            let full_command = format!("/{}", parsed.command);
-
-            if !Self::is_builtin_agent_command(parsed.command) {
-                if let Some(recipe_path) =
-                    crate::slash_commands::recipe_slash_command::get_recipe_for_command(
-                        &full_command,
-                    )
-                {
-                    if recipe_path.exists() {
-                        cx.send_notification(SessionNotification::new(
-                            args.session_id.clone(),
-                            SessionUpdate::AgentMessageChunk(ContentChunk::new(
-                                ContentBlock::Text(TextContent::new(format!(
-                                    "Running recipe: {}",
-                                    full_command
-                                ))),
-                            )),
-                        ))?;
-                    }
-                }
-            }
-        }
-
-        let session_config = SessionConfig {
-            id: session_id.clone(),
-            schedule_id: None,
-            max_turns: None,
-            retry_config: None,
-        };
-
-        let mut stream = agent
-            .reply(user_message, session_config, Some(cancel_token.clone()))
-            .await
-            .internal_err_ctx("Error getting agent reply")?;
-
-        let mut was_cancelled = false;
-        let mut first_event_logged = false;
-        let mut event_count: u32 = 0;
-        // Streaming chain buffer: tracks consecutive tool requests across
-        // `AgentEvent::Message` events so chains that span multiple rows are
-        // still registered. Sequential tool use (Bedrock/Anthropic) yields
-        // request → response → request → response across separate
-        // assistant/user messages, so tool responses are chain-neutral; only
-        // non-tool content (text, thinking, image, etc.) breaks the run.
-        // Holds `(tool_call_id, message_id_of_owning_row)` in arrival order;
-        // re-registered eagerly each time a request arrives so
-        // `handle_tool_response` finds the chain when subsequent responses
-        // are processed.
-        let mut chain_buffer: Vec<(String, String)> = Vec::new();
-
-        while let Some(event) = stream.next().await {
-            if cancel_token.is_cancelled() {
-                was_cancelled = true;
-                break;
-            }
-            event_count += 1;
-            if !first_event_logged {
-                debug!(
-                    target: "perf",
-                    sid = %sid,
-                    ttft_ms = t_start.elapsed().as_millis() as u64,
-                    "perf: prompt first stream event (time-to-first-token from prompt start)"
-                );
-                first_event_logged = true;
-            }
-
-            match event {
-                Ok(crate::agents::AgentEvent::Message(message)) => {
-                    // Agent persists messages via session_manager.add_message() internally.
-                    let stored_message_id = message.id.clone();
-
-                    let mut sessions = self.sessions.lock().await;
-                    let session = sessions.get_mut(&session_id).ok_or_else(|| {
-                        agent_client_protocol::Error::invalid_params()
-                            .data(format!("Session not found: {}", session_id))
-                    })?;
-
-                    for content_item in &message.content {
-                        match content_item {
-                            MessageContent::ToolRequest(tr) => {
-                                if let Some(msg_id) = stored_message_id.as_deref() {
-                                    chain_buffer.push((tr.id.clone(), msg_id.to_string()));
-                                    // Re-register eagerly so the chain is in
-                                    // place by the time the matching
-                                    // `tool_response` triggers
-                                    // `maybe_summarize_chain` (sequential
-                                    // tool use interleaves request/response
-                                    // events).
-                                    extend_chain_membership(
-                                        &chain_buffer,
-                                        &mut session.chain_membership,
-                                    );
-                                }
-                            }
-                            MessageContent::ToolResponse(_) => {
-                                // Chain-neutral: a response between two
-                                // requests doesn't break the run, matching
-                                // the frontend's `groupContentSections`.
-                            }
-                            _ => {
-                                // Text, thinking, image, etc. end the run.
-                                chain_buffer.clear();
-                            }
-                        }
-
-                        self.handle_message_content(
-                            content_item,
-                            &args.session_id,
-                            &session_id,
-                            stored_message_id.as_deref(),
-                            &agent,
-                            session,
-                            cx,
-                        )
-                        .await?;
-                    }
-                }
-                Ok(_) => {}
-                Err(e) => {
-                    return Err(agent_client_protocol::Error::internal_error()
-                        .data(format!("Error in agent response stream: {}", e)));
-                }
-            }
-        }
-
-        {
-            let mut sessions = self.sessions.lock().await;
-            if let Some(session) = sessions.get_mut(&session_id) {
-                // Final safety net: in case the stream ended without any
-                // chain-breaking content, make sure a multi-tool buffer is
-                // registered. (Eager registration during the loop usually
-                // covers this.)
-                extend_chain_membership(&chain_buffer, &mut session.chain_membership);
-                session.cancel_token = None;
-            }
-        }
-
-        let session = self
-            .session_manager
-            .get_session(&session_id, false)
-            .await
-            .internal_err_ctx("Failed to load session")?;
-        let provider = agent
-            .provider()
-            .await
-            .internal_err_ctx("Failed to get provider")?;
-        let usage_update =
-            build_usage_update(&session, provider.get_model_config().context_limit());
-        cx.send_notification(SessionNotification::new(
-            args.session_id.clone(),
-            SessionUpdate::UsageUpdate(usage_update),
-        ))?;
-
-        debug!(
-            target: "perf",
-            sid = %sid,
-            ms = t_start.elapsed().as_millis() as u64,
-            events = event_count,
-            cancelled = was_cancelled,
-            "perf: prompt done"
-        );
-        let stop_reason = if was_cancelled {
-            StopReason::Cancelled
-        } else {
-            StopReason::EndTurn
-        };
-
-        let mut response = PromptResponse::new(stop_reason);
-        if let Some(usage) = build_prompt_usage(&session) {
-            response = response.usage(usage);
-        }
-        Ok(response)
-    }
-
-    async fn on_cancel(
-        &self,
-        args: CancelNotification,
-    ) -> Result<(), agent_client_protocol::Error> {
-        debug!(?args, "cancel request");
-
-        let session_id = args.session_id.0.to_string();
-        let mut sessions = self.sessions.lock().await;
-
-        if let Some(session) = sessions.get_mut(&session_id) {
-            if let Some(ref token) = session.cancel_token {
-                info!(session_id = %session_id, "prompt cancelled");
-                token.cancel();
-            }
-        } else {
-            warn!(session_id = %session_id, "cancel request for unknown session");
-        }
-
-        Ok(())
-    }
-
-    async fn on_set_model(
-        &self,
-        session_id: &str,
-        model_id: &str,
-    ) -> Result<SetSessionModelResponse, agent_client_protocol::Error> {
-        let config = self.config()?;
-        let agent = self.get_session_agent_provider_ready(session_id).await?;
-        let current_provider = agent
-            .provider()
-            .await
-            .internal_err_ctx("Failed to get provider")?;
-        let provider_name = current_provider.get_name().to_string();
-        let current_model_config = current_provider.get_model_config();
-        let extensions =
-            EnabledExtensionsState::for_session(&self.session_manager, session_id, &config).await;
-        let model_config = crate::model::ModelConfig::new(model_id)
-            .invalid_params_err_ctx("Invalid model config")?
-            .with_canonical_limits(&provider_name);
-        let model_config =
-            with_preserved_session_request_params(model_config, Some(&current_model_config), None);
-        let session = self
-            .session_manager
-            .get_session(session_id, false)
-            .await
-            .internal_err_ctx("Failed to get session")?;
-        let provider = self
-            .create_provider(
-                &provider_name,
-                model_config,
-                extensions,
-                Some(session.working_dir),
-            )
-            .await
-            .internal_err_ctx("Failed to create provider")?;
-        agent
-            .update_provider(provider, session_id)
-            .await
-            .internal_err_ctx("Failed to update provider")?;
-        let mode = agent.goose_mode().await;
-        agent
-            .update_goose_mode(mode, session_id)
-            .await
-            .internal_err_ctx("Failed to propagate mode")?;
-        // model_config is already updated on the session by the agent's update_provider call.
-        Ok(SetSessionModelResponse::new())
-    }
-
-    async fn build_config_update(
-        &self,
-        session_id: &SessionId,
-    ) -> Result<(SessionNotification, Vec<SessionConfigOption>), agent_client_protocol::Error> {
-        let session = self
-            .session_manager
-            .get_session(&session_id.0, false)
-            .await
-            .internal_err()?;
-        let agent = self.get_session_agent_provider_ready(&session_id.0).await?;
-        let provider = agent
-            .provider()
-            .await
-            .internal_err_ctx("Failed to get provider")?;
-        let provider_name = provider.get_name().to_string();
-        let current_model = provider.get_model_config().model_name.clone();
-        let goose_mode = agent.goose_mode().await;
-        let inventory = self
-            .provider_inventory
-            .entry_for_provider(&provider_name)
-            .await
-            .internal_err()?;
-        let Some(inventory) = inventory else {
-            return Err(agent_client_protocol::Error::internal_error()
-                .data(format!("Unknown provider inventory: {}", provider_name)));
-        };
-        let model_state = build_model_state(current_model.as_str(), &inventory);
-        let mode_state = build_mode_state(goose_mode)?;
-        let provider_options = build_provider_options(Some(&provider_name)).await;
-        let config_options = build_config_options(
-            &mode_state,
-            &model_state,
-            session_provider_selection(&session),
-            provider_options,
-        );
-        let notification = SessionNotification::new(
-            session_id.clone(),
-            SessionUpdate::ConfigOptionUpdate(ConfigOptionUpdate::new(config_options.clone())),
-        );
-        Ok((notification, config_options))
-    }
-
-    async fn on_set_mode(
-        &self,
-        session_id: &str,
-        mode_id: &str,
-    ) -> Result<SetSessionModeResponse, agent_client_protocol::Error> {
-        let mode = mode_id.parse::<GooseMode>().map_err(|_| {
-            agent_client_protocol::Error::invalid_params()
-                .data(format!("Invalid mode: {}", mode_id))
-        })?;
-
-        let agent = self.get_session_agent_provider_ready(session_id).await?;
-        agent
-            .update_goose_mode(mode, session_id)
-            .await
-            .internal_err_ctx("Failed to update mode")?;
-
-        // goose_mode is already updated on the session above.
-
-        Ok(SetSessionModeResponse::new())
-    }
-
-    async fn update_provider(
-        &self,
-        session_id: &str,
-        provider_name: &str,
-        model_name: Option<&str>,
-        context_limit: Option<usize>,
-        request_params: Option<std::collections::HashMap<String, serde_json::Value>>,
-    ) -> Result<(), agent_client_protocol::Error> {
-        let config = self.config()?;
-        let agent = self.get_session_agent_provider_ready(session_id).await?;
-        let current_provider = agent
-            .provider()
-            .await
-            .internal_err_ctx("Failed to get provider")?;
-        let current_provider_name = current_provider.get_name();
-        let current_model_config = current_provider.get_model_config();
-        let current_model = current_model_config.model_name.clone();
-        let has_default_overrides =
-            model_name.is_some() || context_limit.is_some() || request_params.is_some();
-        let use_default_provider = provider_name == DEFAULT_PROVIDER_ID;
-        let resolved_provider_name = if use_default_provider {
-            config
-                .get_goose_provider()
-                .internal_err_ctx("Failed to resolve default provider from config")?
-        } else {
-            provider_name.to_string()
-        };
-        let is_changing_provider = resolved_provider_name != current_provider_name;
-        let default_model = if let Some(model_name) = model_name {
-            model_name.to_string()
-        } else if use_default_provider {
-            config
-                .get_goose_model()
-                .internal_err_ctx("Failed to resolve default model from config")?
-        } else if is_changing_provider {
-            ACP_CURRENT_MODEL.to_string()
-        } else {
-            current_model
-        };
-        let model = model_name.unwrap_or(&default_model);
-        let mut model_config = crate::model::ModelConfig::new(model)
-            .invalid_params_err_ctx("Invalid model config")?
-            .with_canonical_limits(&resolved_provider_name)
-            .with_context_limit(context_limit);
-        model_config = with_preserved_session_request_params(
-            model_config,
-            (!is_changing_provider).then_some(&current_model_config),
-            request_params,
-        );
-
-        let extensions =
-            EnabledExtensionsState::for_session(&self.session_manager, session_id, &config).await;
-        let session = self
-            .session_manager
-            .get_session(session_id, false)
-            .await
-            .internal_err_ctx("Failed to get session")?;
-        let new_provider = self
-            .create_provider(
-                &resolved_provider_name,
-                model_config,
-                extensions,
-                Some(session.working_dir),
-            )
-            .await
-            .internal_err_ctx("Failed to create provider")?;
-        agent
-            .update_provider(new_provider, session_id)
-            .await
-            .internal_err_ctx("Failed to update provider")?;
-        let mode = agent.goose_mode().await;
-        agent
-            .update_goose_mode(mode, session_id)
-            .await
-            .internal_err_ctx("Failed to propagate mode")?;
-        let provider = agent
-            .provider()
-            .await
-            .internal_err_ctx("Failed to get provider")?;
-
-        // provider_name is already updated on the session by the agent's update_provider call.
-
-        if use_default_provider {
-            let update = self
-                .session_manager
-                .update(session_id)
-                .provider_name(DEFAULT_PROVIDER_ID);
-            if has_default_overrides {
-                update
-                    .model_config(provider.get_model_config())
-                    .apply()
-                    .await
-                    .internal_err_ctx("Failed to persist default provider selection overrides")?;
-            } else {
-                update
-                    .clear_model_config()
-                    .apply()
-                    .await
-                    .internal_err_ctx("Failed to persist default provider selection")?;
-            }
-        }
-        Ok(())
-    }
-
-    async fn on_list_sessions(
-        &self,
-        req: ListSessionsRequest,
-    ) -> Result<ListSessionsResponse, agent_client_protocol::Error> {
-        if let Some(cwd) = req.cwd.as_deref() {
-            if !cwd.is_absolute() {
-                return Err(agent_client_protocol::Error::invalid_params()
-                    .data("cwd must be an absolute path"));
-            }
-        }
-
-        let cwd = req.cwd.as_deref();
-        let cursor =
-            decode_session_list_cursor(req.cursor.as_deref(), cwd, &ACP_SESSION_LIST_TYPES)?;
-
-        // ACP clients see their own (Acp) sessions plus legacy User/Scheduled ones.
-        let page = self
-            .session_manager
-            .list_nonempty_sessions_by_types_paged(
-                &ACP_SESSION_LIST_TYPES,
-                cwd,
-                cursor.as_ref(),
-                SESSION_LIST_PAGE_SIZE,
-            )
-            .await
-            .internal_err()?;
-        let session_infos: Vec<SessionInfo> = page
-            .sessions
-            .into_iter()
-            .map(|s| {
-                let meta = session_meta(&s);
-                SessionInfo::new(SessionId::new(s.id), s.working_dir)
-                    .title(s.name)
-                    .updated_at(s.updated_at.to_rfc3339())
-                    .meta(meta)
-            })
-            .collect();
-        let next_cursor = page
-            .next_cursor
-            .as_ref()
-            .map(|cursor| encode_session_list_cursor(cursor, cwd, &ACP_SESSION_LIST_TYPES))
-            .transpose()?;
-        Ok(ListSessionsResponse::new(session_infos).next_cursor(next_cursor))
-    }
-
-    async fn on_fork_session(
-        &self,
-        cx: &ConnectionTo<Client>,
-        args: ForkSessionRequest,
-    ) -> Result<ForkSessionResponse, agent_client_protocol::Error> {
-        validate_absolute_cwd(&args.cwd)?;
-        let source_session_id = &*args.session_id.0;
-
-        let new_session = self
-            .session_manager
-            .copy_session(source_session_id, "Fork".to_string())
-            .await
-            .internal_err()?;
-        let new_session_id = new_session.id.clone();
-
-        // Update working dir for the fork.
-        self.session_manager
-            .update(&new_session_id)
-            .working_dir(args.cwd.clone())
-            .apply()
-            .await
-            .internal_err()?;
-
-        let goose_session = self
-            .session_manager
-            .get_session(&new_session_id, false)
-            .await
-            .internal_err()?;
-
-        let (agent_tx, agent_rx) = tokio::sync::watch::channel::<AgentSetupSignal>(None);
-
-        let acp_session = GooseAcpSession {
-            agent: AgentHandle::Loading(agent_rx),
-            tool_requests: HashMap::new(),
-            chain_membership: HashMap::new(),
-            responded_tool_ids: HashSet::new(),
-            summarized_chains: HashSet::new(),
-            cancel_token: None,
-            pending_working_dir: None,
-        };
-        self.sessions
-            .lock()
-            .await
-            .insert(new_session_id.clone(), acp_session);
-
-        let mode_state = build_mode_state(self.goose_mode)?;
-        let resolved = resolve_provider_and_model(&self.config_dir, &goose_session).await;
-        let (model_state, config_options, prebuilt_provider) = self
-            .prepare_session_init_config(&resolved, &mode_state, &goose_session)
-            .await;
-
-        let acp_session_id = SessionId::new(new_session_id.clone());
-
-        self.spawn_agent_setup(
-            cx,
-            agent_tx,
-            AgentSetupRequest {
-                session_id: acp_session_id.clone(),
-                goose_session,
-                mcp_servers: args.mcp_servers,
-                resolved_provider: resolved.ok(),
-                prebuilt_provider,
-            },
-        );
-
-        let meta = session_meta(&new_session);
-
-        let mut response = ForkSessionResponse::new(acp_session_id.clone())
-            .modes(mode_state)
-            .meta(meta);
-
-        if let Some(ms) = model_state {
-            response = response.models(ms);
-        }
-        if let Some(co) = config_options {
-            response = response.config_options(co);
-        }
-        Self::send_available_commands_update(cx, &acp_session_id, &args.cwd)?;
-        Ok(response)
-    }
-
-    async fn on_close_session(
-        &self,
-        session_id: &str,
-    ) -> Result<CloseSessionResponse, agent_client_protocol::Error> {
-        let mut sessions = self.sessions.lock().await;
-        if let Some(session) = sessions.get(session_id) {
-            if let Some(ref token) = session.cancel_token {
-                token.cancel();
-            }
-        }
-        sessions.remove(session_id);
-        info!(session_id = %session_id, "ACP session closed");
-        Ok(CloseSessionResponse::new())
-    }
 }
 
 pub struct GooseAcpHandler {
